pyproject.toml
[project]
name = "ghost-writer-mcp"
version = "0.1.0"
description = "MCP server that turns GitHub activity into LinkedIn thought leadership"
license = {text = "MIT"}
keywords = ["mcp", "linkedin", "github", "ai", "llm", "content-generation"]
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0",
    "PyGithub>=2.0",
    "openai>=1.0",
    "anthropic>=0.42",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.scripts]
ghost-writer-mcp = "src.server:main"

[project.optional-dependencies]
dev = ["pytest>=9.0", "ruff>=0.14"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
